+++ libft/ft_calloc.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_calloc.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hkasamat <hkasamat@student.42.fr>          +#+  +:+       +#+        */
+/*   By: haruki <haruki@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/09 15:14:05 by hkasamat          #+#    #+#             */
-/*   Updated: 2024/11/30 16:29:50 by hkasamat         ###   ########.fr       */
+/*   Updated: 2024/12/09 17:12:51 by haruki           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
+#include "libft.h"
 
-void	*ft_calloc(unsigned int nmemb, unsigned int size)
+void	*ft_calloc(size_t nmemb, size_t size)
 {
-	void			*ptr;
-	unsigned int	i;
+	void	*ptr;
+	size_t	i;
 
 	i = 0;
 	if (size == 0 || nmemb == 0)
